--- Source/Iridio/Common/Utils/StringUtils.cs
@@ -1,15 +1,0 @@
-﻿using System.Collections.Generic;
-using System.IO;
-
-namespace Iridio.Common.Utils
-{
-    public static class StringUtils
-    {
-        public static IEnumerable<string> Lines(this string s)
-        {
-            using var tr = new StringReader(s);
-            while (tr.ReadLine() is { } l)
-                yield return l;
-        }
-    }
-}
