+++ Source/Iridio/Preprocessing/IDirectoryContext.cs
@@ -1,0 +1,7 @@
+﻿namespace Iridio.Preprocessing
+{
+    public interface IDirectoryContext
+    {
+        string WorkingDirectory { get; set; }
+    }
+}
