+++ Source/Iridio/Preprocessing/ITextFileFactory.cs
@@ -1,0 +1,7 @@
+﻿namespace Iridio.Preprocessing
+{
+    public interface ITextFileFactory
+    {
+        ITextFile Get(string path);
+    }
+}
