+++ Source/Iridio/Preprocessing/TextFileFactory.cs
@@ -1,0 +1,10 @@
+﻿namespace Iridio.Preprocessing
+{
+    public class TextFileFactory : ITextFileFactory
+    {
+        public ITextFile Get(string path)
+        {
+            return new TextFile(path);
+        }
+    }
+}
